requirements.txt
scikit-learn
numpy
pillow
opencv-python
joblib
pandas
scikit-image
